POC/pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "final-project"
authors = [
    {name = "aditi"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "streamlit-code-editor (>=0.1.22,<0.2.0)",
    "leetscrape (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
